requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.9.1
httpx==0.27.0
pyjwt==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
